feat(app-config): extract shared @vt/app-config/folders module
Folder FS operations (directory scanning, folder-tree sidebar payload, the
add-folder selector, folder mutations, starred folders) move into a single
@vt/app-config/folders module so VTD's browser-mode gateway and the Electron
main process build identical results from one source. Electron call sites
(folderScanning, watchFolder, daemon-folder-tree-sync) become thin shims/
adapters over it — no behaviour change, duplication removed.
