+++ packages/libraries/app-config/package.json
@@ -8,6 +8,7 @@
     "./settings": "./src/settings/settings_IO.ts",
     "./project-config": "./src/project-config/voicetree-config-io.ts",
     "./project": "./src/project/index.ts",
+    "./folders": "./src/folders/index.ts",
     "./node-layout-io": "./src/positions/node-layout-io.ts"
   },
   "dependencies": {
